+++ rosie/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this plugin will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this plugin adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.9] - 2026-01-20
+
+### Changed
+- Serializer scripts are now added via hook instead of via the `use-unity-editor` skill.
+- Serializer scripts are now added to the project's `Assets/Editor/` folder instead of the `Assets/Scripts/Editor/` folder.
+
 ## [0.4.8] - 2026-01-09
 
 ### Changed